Adding more exercises

+++ 02_Exercises/01_slice/main.go
@@ -2,8 +2,8 @@ package main
 
 import "fmt"
 
-func half(x int) (float64, bool) {
-	return float64(x) / 2, x%2 == 0
+func half(x float64) (float64, bool) {
+	return x / 2, x%2.0 == 0
 }
 
 func main() {
